+++ realm/realm-library/src/androidTestObjectServer/java/io/realm/SyncConfigurationTests.java
@@ -35,8 +35,6 @@ import java.util.Map;
 
 import io.realm.entities.AllJavaTypes;
 import io.realm.entities.StringOnly;
-import io.realm.log.LogLevel;
-import io.realm.log.RealmLog;
 import io.realm.rule.RunInLooperThread;
 import io.realm.rule.TestSyncConfigurationFactory;
 
@@ -394,8 +392,6 @@ public class SyncConfigurationTests {
 
     @Test
     public void initialData() {
-        RealmLog.clear();
-        RealmLog.setLevel(LogLevel.ALL);
         SyncUser user = createTestUser();
         String url = "realm://objectserver.realm.io/default";